reorganize query planning and execution
